requirements_JoyTTS.txt
absl-py==2.2.2
accelerate==1.7.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
airportsdata==20250224
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
beautifulsoup4==4.13.3
blake3==1.0.4
boto3==1.38.0
botocore==1.38.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cn2an==0.5.23
coloredlogs==15.0.1
compressed-tensors==0.9.1
conformer==0.3.2
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cupy-cuda12x==13.4.1
cycler==0.12.1
Cython==3.0.12
datasets==2.18.0
decorator==5.2.1
deepspeed==0.14.2
depyf==0.18.0
diffusers==0.32.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
editdistance==0.8.1
einops==0.8.1
einx==0.3.0
email_validator==2.2.0
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
fastrlock==0.8.3
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.2.0
funasr==1.2.6
gast==0.6.0
gdown==5.1.0
gguf==0.10.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.70.0
gradio==5.4.0
gradio_client==1.4.2
groovy==0.1.2
grpcio==1.57.0
grpcio-tools==1.57.0
h11==0.14.0
hf-xet==1.1.3
hjson==3.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
ijson==3.3.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
inflect==7.3.1
iniconfig==2.1.0
interegular==0.3.3
ipdb==0.13.13
ipython==8.35.0
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jiwer==3.1.0
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaldiio==2.18.1
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
librosa==0.10.2
lightning==2.5.0.post0
lightning-utilities==0.14.3
llguidance==0.7.29
llvmlite==0.43.0
lm-format-enforcer==0.10.11
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.5.4
modelscope==1.15.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
onnxruntime-gpu==1.19.0
openai==1.74.0
openai-whisper==20240930
opencv-python-headless==4.11.0.86
opentelemetry-api==1.34.0
opentelemetry-exporter-otlp==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-exporter-otlp-proto-http==1.34.0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
opentelemetry-semantic-conventions-ai==0.4.9
orjson==3.10.16
oss2==2.19.1
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post5
peft==0.15.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
proces==0.1.7
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==4.25.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pynvml==12.0.0
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.12
pytorch-lightning==2.5.1
pytorch-wpe==0.0.1
pytz==2025.2
pyworld==0.3.4
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
ray==2.40.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rich-toolkit==0.14.1
rpds-py==0.24.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.11.5
s3transfer==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorrt-cu12==10.0.1
tensorrt-cu12-bindings==10.0.1
tensorrt-cu12-libs==10.0.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.5.1
torch-complex==0.4.4
torchaudio==2.5.1
torchmetrics==1.7.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.1.0
typeguard==4.4.2
typer==0.15.2
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
uvicorn==0.30.0
uvloop==0.21.0
vector-quantize-pytorch==1.22.15
vllm==0.7.3
vocos==0.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
WeTextProcessing==1.0.3
wget==3.2
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
yapf==0.43.0
yarl==1.19.0
zhconv==1.4.3
zhon==2.1.1
zipp==3.21.0
